Fix for issue #18 (#26)
We create new Kpoints classes to read and refine the data from the HDF5 file. This removes some logic from the Band class that more logically belongs to the Kpoints class. In the process, we change the logic such that different input modes for different KPOINTS files do not crash the Band class.