Cleanup NGL Viewer class (#11)
A lot of the functionality previously contained in the Structure class
is now moved to the Viewer3d class. The main reason for this change is
that the structure class does not need to now about NGL at all.
Also the interface is much leaner and more options can be choosen by the
user afterwards if desired and don't expand the default interface.