Changed the implementation so that also non-orthogonal z-axis are allowed.
Also made sure that strange fractional coordinates (e.g. -2.234) are correctly mapped between 0 and 1.

For constant current mode, the smoothed charge density is now also rolled in a way that values on both sides of the unit cell boarders are possible.

Tests were updated, but I still need to double check if the STM pictures are actually correct in some edge cases.